update Tuya IoTOS 2.3.2

+++ sdk/include/crc32i.h
@@ -21,7 +21,7 @@
 #define __CR_CRC32I_H__ 
 
 unsigned int hash_crc32i_init (void);
-unsigned int hash_crc32i_update (unsigned int hash,const void *data,unsigned short size);
+unsigned int hash_crc32i_update (unsigned int hash,const void *data,unsigned int size);
 unsigned int hash_crc32i_finish (unsigned int hash);
 
 /*
@@ -29,7 +29,7 @@ unsigned int hash_crc32i_finish (unsigned int hash);
     CRC-32/IEEE 一次性完成
 =======================================
 */
-unsigned int hash_crc32i_total (const void* data,unsigned short size);
+unsigned int hash_crc32i_total (const void* data,unsigned int size);
 
 #endif
